requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
jinja2>=3.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# MySQL
pymysql>=1.1.0
cryptography>=41.0.0

# PostgreSQL
psycopg2-binary>=2.9.0

# SQLite (built-in with Python, no extra package needed)

